Basic v4l2 driver for MobirAir in Python
This is a very basic version of driver that reads frames from the
MobirAir thermal camera and dumps it into a v4l2 loopback device.

Initial setup require the v4l2loopback module to be loaded with at least
one camera device (the path is currently hard coded - sorry)

This driver implements the minimal necessary initialization routine to
get (usable) raw data from the camera. It currently does not do anything
useful with during the calibration routine or some fancy thermal
processing.

Even the header - which is in every frame - is not really being parsed.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>