Allow disabling USB data retrieve thread
The thread is responsible for image processing and shouldn't be burdened with
much else. As the camera uses the same channel for the stream as it does
in order to answer to data requests (calibration files, temperature
mappings, …), this thread needs to be toggled.
This way we can retrieve some data synchronously (with disabled thread),
while mostly running asynchron.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>