Implement `getAllKData` using new `GetArmParam`
The MobirAir speaks a bit of an interesting protocol. Especially
retrieval of calibration data happens through a (limited) command that
directly dumps the flash memory from the specified location.

This command (`GetArmParam`) is now implemented. Also note that most of
the protocol has been wrapped in the `MobirAirUSBProtocol` class.
e.g. using `GetArmParam` becomes easier, as the protocol class deals
with the limitation on returned byte length per call, the Shutter On and
Off are more understandable now (original naming is inverted), …

A bit of a problem currently is, who should call `getAllKData` (and
similar methods). Is it the responsibility of driver to provide all
required data for further processing or should the classes handling them
be made responsible?

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>