fix: allowing reuse of camera after script restart
Previously the MobirAir camera would frequently not send anything after
the script was restarted and an `usb.reset()` had been triggered. As it
turns out this seemed to be due to a missing shut down process, in which
e.g. the stream, buffer, … weren't correctly handled.

This fix improves on this issue and allows for reliable reuse of the
camera when developing the script.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>