Add doNUCbyTwoPoint image correction
Doing a NUC is relatively important for thermal cameras as the raw
data often contains heavy static noise and temperature distortions.

The doNUCbyTwoPoint is implemented as by the MobirAir app itself and
uses a combination of calibration data generated at manufacturing
(temperature distortions) and during a shutter sequence in order to
correct the live images.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>