Add changeR method
The original driver has a changeR method that is triggered every 25
frames.

It seems to adjust the raw data processing (by selecting the right k
data array) and camera settings (using `SetDetectIndex`) according to
the current temperature by the fpa.

The FPA temperature is included in the parameter line of every frame,
but in order to make it useable, it needs a bit of preprocessing (see
`FixedParamLine.getRealtimeFpaTemp`).

Most of the concepts for the algorithms and preprocessing routines are
from the original MobirAir android app.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>