fix: reduce image tearing by increasing read size
When using a too small read buffer for retrieving thermal frames,
reduces in tearing of the images.

The size was previously chosen quite conservatively as it otherwise
could result in USB Overflow errors. But as it turns out, increasing it
significantly is quite important to improve read speeds and doesn't
result in a crash, if the size has certain constraints (2ⁿ number)

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>