Add `calcMeasureKj`
Method is triggered after each calibration step and seems to be
responsible for setting the `kj` parameter.

In this step, we also introduce a temperature utility class that is
responsible for converting temperatures. In our case right now, it
converts the values from the frame header into °C.

Note here that it has been decided to use the temperatures mostly as
floats in the code. This is in contrast to the MobirAir app, which often
converts the floats into shorts by multiplying them by 100.
So be aware of that.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>