Implement CustomParamLine
`CustomParamLine` is quite important later on for converting the raw
data into usable temperatures. It is part of the header in each frame.

It's interesting to note here how the bit fields are read because it is
slightly weird. The header is first parsed as a short array, from which
single bytes are then read using bit fields.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>