state: Moved param into MeasureParams
A large block of the data that is later on important for temperature
calculations are included in the frames. Instead of using the
`lastFrame` attribute of the state, we introduce a `MeasureParams` data
class that holds the important information.

This also streamlines the process for last shutter data and more.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>