update measure parameters with every frame
The original app only updates them with every 25th frame. This
essentially reduces to an update frequency of 1Hz for this data block.

This is analog to the behavior of the app: measurement points and
temperature calculations only update about every second. As we want to
update the temperatures much more frequently, it makes sense to update
the data class according with every frame instead.

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>