Add first temperature calculations
By using numpy and the methods already explored in the to_temperature
image analysis scripts, it became feasible to transform all raw frames
into temperature frames. The methods in MobileAirTempUtil are already
tested and seem to work correctly.

But as of right now, there seem to be a few mistakes somewhere. The
temps are quite off from the reality, so further debugging is needed.
But hey - We have temps

Signed-off-by: Sophie Tyalie <dev@flowerpot.me>